pyproject.toml
# -----------------------------------------------------------------------------
# Build System Configurations
# -----------------------------------------------------------------------------

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

# -----------------------------------------------------------------------------
# Project Configurations
# -----------------------------------------------------------------------------
#
[project]
name = "socx"
version = "0.1.0"
readme = "README.md"
license = "Apache-2.0"
authors = [ { name = "Sagi Kimhi", email = "sagi.kim5@gmail.com" } ]
maintainers = [ { name = "Sagi Kimhi", email = "sagi.kim5@gmail.com" } ]
description = "Tool executer and plugin manager for SoC team."
dependencies = [
    "click>=8.1.8",
    "rich-cli>=1.8.0",
    "python-dotenv[cli]>=1.0.1",
    "dynaconf>=3.2.7",
    "jinja2>=3.1.5",
    "flask>=3.1.0",
    "textual-dev>=1.0.0",
    "textual>=0.1.18",
]
license-files = ["LICENSE"]
requires-python = ">=3.12"

[project.scripts]
socx='socx.__main__:cli'

[project.gui-scripts]

[project.optional-dependencies]
dev = [
    "mypy",
    "ruff",
    "pytest",
    "pytest-cov",
    "pyupgrade",
    "pydocstyle",
    "pycodestyle",
    "flake8-bugbear",
    "flake8-simplify",
    "debugpy>=1.8.12",
]

# -----------------------------------------------------------------------------
# Developer Configurations
# -----------------------------------------------------------------------------

[tool.mypy]
files = [ "src" ]
python_version = "3.12"
pretty = true
strict = true
color_output = true
show_error_codes = true
warn_unreachable = true

